feat: make tiles 64x32 (#28)
The tiles are now 64x32 instead of 256x128. Blender still essentially outputs 256x128 tiles, but the sprite sheet stitch process performed via imagemagik scales them down to 64x128. The Blender files also required modification so that the tile edges would actually align.

In Blender, I've 'zoomed' the camera in. The consequence is that the Northern, Eastern, Southern and Western edges get clipped, but this does ensure that the diagonal edges align perfectly when rendered. I think diagonal edges in images have a rounding error factor because of the dithering and the more you scale down, the bigger the problem becomes. I'm not completely happy with this, but I think it'll hold in the short to medium term. A better solution might be for the tile loader to 'clip' perform the zoom/clip. That way the tile scale size won't matter to Blender. Another idea might be to find a way to calculate an overlap offset accurately when doing tile placement, I found that I could get perfect alignment by adding magic offset values, but I spent a lot time trying to eliminate magic offsets and I don't intend on letting that crap back in. I don't mind if offsets determined by straightforward formulas that can be explained easily and proven not to be hacks.

I also adjusted the desert tile shader for aesthetics and the zoom out limit (more sprites = higher GPU load, I've picked values that maintain an easy 60FPS).

Jira: https://bluestickerclub.atlassian.net/jira/software/projects/ROV/boards/1?selectedIssue=ROV-12